# Prevent empty buffers from being created for debug data when unused