We currently create a a lot of empty buffers that are placeholder for debug data, which end up being unused for the regular use case. I have not seen any guarding against 0-sized buffer using gl.bufferData in ANGLE or the Chromium source code. In a crowded label areas this can easily reach ~1000 empty vertex buffer handles after panning for a few seconds.

<details>
<summary>Profiling data</summary>

For one map load on Chrome, the observed time taken for the creation of these empty buffers over the default area of Washington (zoom level 12 for the debug page).

In `symbol_bucket.js` `SymbolBucket.constructor`, creation of object textCollisionBox, iconCollisionBox, textCollisionCircle, iconCollisionCircle

```
2.8Ghz i7 (macbook 2019)
Average: 27.14285771932681 us
Total: 1330.0000282470137 us

2.4Ghz i5 (macbook 2011)
Average: 140.6338028093006 us
Total: 9984.999999460342 us

Galaxy s6 (smaller viewport)
Average: 195.83333333865008 us
Total: 4700.000000127602 us
```

In `symbol_bucket.js` `SymbolBucket.upload`, upload for textCollisionBox, iconCollisionBox, textCollisionCircle, iconCollisionCircle:

```
2.8Ghz i7 (macbook 2019)
Average: 78.65079290543993 us
Total: 4954.999953042716 us

2.4Ghz i5 (macbook 2011)
Average: 171.74604175009377 us
Total: 10819.999999625097 us

Galaxy s6 (smaller viewport)
Average: 191.66666666471124 us
Total: 4599.99999995307 us
```
Refer to 80572680668fb85a6577a43ab38b57e339fc6b02
</details>

Deferring the creation at the time of use for these buffers (`map.showCollisionBoxes = true`) should show savings between ~1 to ~5ms on the tile workers and ~4.5 to ~11 ms on the main thread for one map load at zoom 12 in washington dc area (which varies depending on the platform). 

## Launch Checklist

<!-- Thanks for the PR! Feel free to add or remove items from the checklist. -->

 - [x] briefly describe the changes in this PR
 - [x] post benchmark scores
 - [x] manually test the debug page
 - [x] @mapbox/gl-native those buffer seem to exist in the native codebase, but I have not checked whether they are created while not being used
 - [x] apply changelog label ('bug', 'feature', 'docs', etc) or use the label 'skip changelog'
 - [x] add an entry inside this element for inclusion in the `mapbox-gl-js` changelog: `<changelog>Prevent empty buffers from being created for debug data when unused</changelog>`